Move `TabletManagerClient` mock to `tmclient/mock` package
The gomock-generated `MockTabletManagerClient` lived in
`go/vt/vtorc/logic/mock_tmclient_test.go`, making it only usable
within that single package's tests. This moves it to
`go/vt/vttablet/tmclient/mock/` as a shared package so any test
can import it.
