+++ go/vt/vttablet/tmclient/mock/client.go
@@ -3,11 +3,11 @@
 //
 // Generated by this command:
 //
-//	mockgen -destination go/vt/vtorc/logic/mock_tmclient_test.go -package logic vitess.io/vitess/go/vt/vttablet/tmclient TabletManagerClient
+//	mockgen -destination go/vt/vttablet/tmclient/mock/client.go -package mock vitess.io/vitess/go/vt/vttablet/tmclient TabletManagerClient
 //
 
-// Package logic is a generated GoMock package.
-package logic
+// Package mock is a generated GoMock package.
+package mock
 
 import (
 	context "context"
